+++ type-tests/tsrx.test-d.ts
@@ -1,4 +1,7 @@
-import sunfallArcTsrxDefault, { sunfallArcTsrx, type SunfallArcTsrxOptions } from "@sunfall/arc-tsrx";
+import sunfallArcTsrxDefault, {
+  sunfallArcTsrx,
+  type SunfallArcTsrxOptions,
+} from "@sunfall/arc-tsrx";
 
 const tsrxExports: Array<unknown> = [sunfallArcTsrxDefault, sunfallArcTsrx];
 type TsrxOptions = SunfallArcTsrxOptions;